+++ sctoolbox/tools/qc_filter.py
@@ -194,7 +194,7 @@ def predict_cell_cycle(adata, species, s_genes=None, g2m_genes=None, inplace=Tru
 
 @deco.log_anndata
 def estimate_doublets(adata, threshold=0.25, inplace=True, plot=True,
-                      groupby=None, threads=4, **kwargs) -> Optional[anndata.AnnData]:
+                      groupby=None, threads=4, fill_na=True, **kwargs) -> Optional[anndata.AnnData]:
     """
     Estimate doublet cells using scrublet.
 
@@ -212,11 +212,12 @@ def estimate_doublets(adata, threshold=0.25, inplace=True, plot=True,
     plot : bool, default True
         Whether to plot the doublet score distribution.
     groupby : str, default None
-        Key in adata.obs to use for batching during doublet estimation.
-        If threads > 1, the adata is split into separate runs across threads.
-        Otherwise each batch is run separately.
-    threads : int, default 4
-        Number of threads used for estimation.
+        Key in adata.obs to use for batching during doublet estimation. If threads > 1,
+        the adata is split into separate runs across threads. Otherwise each batch is run separately.
+    fill_na : bool, default True
+        If True, replaces NA values returned by scrublet with 0 and False. Scrublet returns NA if it cannot calculate
+        a doublet score. Keep in mind that this does not mean that it is no doublet.
+        By setting this parameter true it is assmuned that it is no doublet.
     **kwargs :
         Additional arguments are passed to scanpy.external.pp.scrublet.
 
@@ -287,6 +288,14 @@ def estimate_doublets(adata, threshold=0.25, inplace=True, plot=True,
     adata.obs["predicted_doublet"] = obs_table["predicted_doublet"]
     adata.uns["scrublet"] = uns_dict
 
+    if fill_na:
+        adata.obs[["doublet_score", "predicted_doublet"]] = (
+            utils.fill_na(adata.obs[["doublet_score", "predicted_doublet"]], inplace=False))
+
+    # Check if all values in colum are of type boolean
+    if adata.obs["predicted_doublet"].dtype != "bool":
+        logger.warning("Could not estimate doublets for every barcode. Columns can contain NAN values.")
+
     # Plot the distribution of scrublet scores
     if plot is True:
         sc.external.pl.scrublet_score_distribution(adata)
@@ -352,6 +361,10 @@ def predict_sex(adata, groupby, gene="Xist",
     save : str, default None
         If provided, the plot will be saved to this path.
 
+    Notes
+    -----
+    adata.X will be converted to numpy.ndarray if it is of type numpy.matrix.
+
     Returns
     -------
     None
@@ -372,7 +385,16 @@ def predict_sex(adata, groupby, gene="Xist",
     if len(gene_index) == 0:
         logger.info("Selected gene is not present in the data. Prediction is skipped.")
         return
-    adata_copy.obs["gene_expr"] = adata_copy.X[:, gene_index].todense().A1
+
+    # If adata.X is of type matrix convert to ndarray
+    if isinstance(adata_copy.X, np.matrix):
+        adata_copy.X = adata_copy.X.getA()
+
+    # Try to flatten for adata.X np.ndarray. If not flatten for scipy sparse matrix
+    try:
+        adata_copy.obs["gene_expr"] = adata_copy.X[:, gene_index].flatten()
+    except AttributeError:
+        adata_copy.obs["gene_expr"] = adata_copy.X[:, gene_index].todense().A1
 
     # Estimate which samples are male/female
     logger.info("Estimating male/female per group")
@@ -945,6 +967,9 @@ def _filter_object(adata, filter, which="obs", remove_bool=True, inplace=True):
         if filter not in table.columns:
             raise ValueError(f"Column {filter} not found in {table_name}.columns")
 
+        if table[filter].dtype.name != "bool":
+            raise ValueError(f"Column {filter} contains values that are not of type boolean")
+
         boolean = table[filter].values
         if remove_bool is True:
             boolean = ~boolean
@@ -955,7 +980,7 @@ def _filter_object(adata, filter, which="obs", remove_bool=True, inplace=True):
         if len(not_found) > 0:
             logger.info(f"{len(not_found)} {element_name} were not found in adata and could therefore not be removed. These genes are: {not_found}")
 
-        boolean = ~table.index.isin(filter).values
+        boolean = ~table.index.isin(filter)
 
     # Remove genes from adata
     if inplace: